duplicity: drop --no-use-agent arg
It's only useful for gpg1 on old Ubuntus (or other distros where
they patched the default to be --use-agent).

Nowadays, I see more people confused by gpg2 saying it's an
obsolete option. So let's remove any worry that there the message
means there is an encryption problem.

Both the snap and flatpak versions ship with gpg2.