Support notifications when the desktop file isn't available
For example, the gnome-shell notification daemon does not know
about the desktop file for either:
- our snap installs (installed with a non-appid filename) nor
- our local dev installs

With this commit, we now support the DEJA_DUP_NOTIFICATION_ICON
environment variable, which we'll use instead of the themed
icon in cases where the installed icon is not in a theme.

And we work around a quirk of the freedesktop notification backend
that doesn't use a default action unless one is set.