Add extremely work-in-progress restic backend
Disabled by default (pass -Denable_restic=true when building to
enable). And even then, the user has to opt-in.

There are some known weak points:
- Need to handle running out of space on storage
- Need to relocate restored files when put in custom folder
- Need lots of testing

But it's possible! I'd like to get this merged into main so that
translations, testing, and incremental improvements can start
happening more easily.

This also adds a borg backend, but that's even more sketchy.
It's only included as a proof of concept.