preferences: add 2-day and 3-day frequency options by default
It's been pointed out that daily & weekly is a bit limiting. So
let's throw 2 and 3 day options in there too.

I'm leery to add a built-in option longer than a week, since
that seems like encouraging bad practices.

(One can always change the value in gsettings to any value.)

Thanks to Ivano for prompting for this change.