main: re-center the overview/welcome status pages
Previously, I had aligned them to the top because of a "settling"
issue where the status page would shift down a second or so after
startup. Seemingly because during startup, the header title would
be marked as visible for a second.

Work around that by only marking the bottom bar visible if it has
an allocated width. Now if the status pages are centered, they
don't shift during startup.