remote: fix bogus network-not-available notification
Since GNetworkMonitor can take a moment to settle and show accurate
network status, initialize it as early as possible during setup.

We did this for the monitor, but not for our main deja-dup process.

This fixes a case where automatic backups would think they weren't
connected for a brief moment, even though they were fine.

Also, withdraw the 'pause' notification when we become unpaused,
so we don't leave a bogus status for the user to see.

Fixes: https://gitlab.gnome.org/World/deja-dup/-/issues/272