duplicity: fix case where we can't restore file permissions/mtime
If we are trying to restore a file from a different user, duplicity
normally tries to fix the uid. But if it fails doing that, it won't
adjust the permissions or mtime either.

And since running as root is not truly a supported mode anyway,
duplicity can never successfully fix the uid. Even if we could, I'm
not sure we want to fix up the uid from a different system.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2081983