oauth: switch to custom uri scheme over localhost redirect
Google is deprecating/removing support for localhost redirects when
authorizing an oauth client. So we need to switch to a custom URI
scheme like 'org.dejadup:' and registering that with the system
in our desktop file.

This has pros and cons in terms of UX. Some browsers (like Firefox)
may prompt "what app do you want to use for this org.dejadup link?"
which is a little gross.

But on the plus side, we now get to skip the "auth complete, close
this tab and go back to Backups" screen and just immediately raise
our window when the auth is done.

Only Google is requiring this, but I also changed Microsoft over
to use the same flow, to simplify the code. (And in case they
decide to make the same change.)

Fixes: https://gitlab.gnome.org/World/deja-dup/-/issues/222