header: don't use 'strict' center policy
This lets the header switcher slide to the left a little bit more
often, rather than shifting the whole thing to the bottom of the
window.

The bottom of the window does not look as good when we have both
the switcher and the action bar from the resume tab visible. So
this keeps the window looking nicer longer, as the width shrinks.