duplicity: bump volsize to 200MB
Before this change, we used 25MB for remote storage locations and
50MB for local ones. This was based off of duplicity using 25MB by
default.

But duplicity switched to 200MB back in 2016(!). So let's follow
suit, finally. Seems battle-tested.

Initially, we'll use 200 for both local and remote locations. We
could bump local even higher, but 200 is a big enough change, let's
stick with that for now.