desktop: set X-Purism-FormFactor to flag mobile support
Now that we work decently well on thin screens, let's mark our
support for mobile form factors using the non-standard
X-Purism-FormFactor for now.

In theory, this same information is in the metainfo file, but
phosh at least still only looks in the desktop file for this.
So let's help it out.