build: add -Dpackagekit=enabled option flag
This enables support for linking with packagekit and installing
runtime dependencies on the fly. Previously, this was detected
at build time based on whether packagekit libraries were installed.

This change just makes it more explicit and also disabled by
default. (I chose 'disabled' rather than 'auto' as the default,
because distros often set -Dauto_features=enabled and this
feature is not encouraged as a UX flow.)

Also, this bumps the required version up to 1.2, in order to drop
the annoying I_KNOW_... build flag definition.