Use rclone backend for duplicity cloud storage
With the deprecation of the pydrive Duplicity backend, rather than
change to the new Google-specific backend, let's just unify our
cloud storage solution between Restic and Duplicity.

This should also make it much easier for us to allow advanced users
to use weird custom cloud locations if we want to expose that.

As part of this, we need to bump the minimum duplicity version to
at least 0.8.12, but while we're doing that, we might as well bump
to 2.0.0 to avoid having to handle pre-2.x CLI option changes.